Split agents.py into separate files
